Conform to best practice example structure
