+++ examples/github_actions_orchestration/pipelines/github_example_pipeline/github_example_pipeline.py
@@ -1,0 +1,23 @@
+#  Copyright (c) ZenML GmbH 2022. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at:
+#
+#       https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+#  or implied. See the License for the specific language governing
+#  permissions and limitations under the License.
+
+from zenml.pipelines import pipeline
+
+
+@pipeline
+def github_example_pipeline(first_step, second_step, third_step):
+    # Link all the steps together
+    first_num = first_step()
+    random_num = second_step()
+    third_step(first_num, random_num)
