+++ examples/github_actions_orchestration/steps/second_step/second_step.py
@@ -1,0 +1,22 @@
+#  Copyright (c) ZenML GmbH 2022. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at:
+#
+#       https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+#  or implied. See the License for the specific language governing
+#  permissions and limitations under the License.
+import random
+
+from zenml.steps import Output, step
+
+
+@step(enable_cache=False)
+def get_random_int() -> Output(random_num=int):
+    """Get a random integer between 0 and 10."""
+    return random.randint(0, 10)
